+++ C++/11 DSA problems in c++ for beginners/2124. Check if All A's Appears Before All B's.cpp
