+++ C++/11 DSA problems in c++ for beginners/Allocate books.cpp
