+++ C++/11 DSA problems in c++ for beginners/Count subarray.cpp
