+++ C++/11 DSA problems in c++ for beginners/Third maximum number.cpp
