+++ C++/11 DSA problems in c++ for beginners/beautiful string.cpp
